Adiciona autenticação do perfil de Professor nas views.

+++ opta_project/opta_app/views/grupo_view.py
@@ -6,13 +6,16 @@ from ..forms import GrupoForm
 
 @login_required
 def mostrar_grupo(request, pk_prof, pk_grupo, template_name='opta_app/detalhe_grupo.html'):
-    grupo = get_object_or_404(Grupo, pk=pk_grupo)
-    projetos = Projeto.objects.filter(grupo_id=pk_grupo)
-    data = {}
-    data['pk_prof'] = pk_prof
-    data['grupo'] = grupo
-    data['projetos'] = projetos
-    return render(request, template_name, data)
+    professor = get_object_or_404(Professor, pk=pk_prof)
+    if request.user.professor == professor:
+        grupo = get_object_or_404(Grupo, pk=pk_grupo)
+        projetos = Projeto.objects.filter(grupo_id=pk_grupo)
+        data = {}
+        data['pk_prof'] = pk_prof
+        data['grupo'] = grupo
+        data['projetos'] = projetos
+        return render(request, template_name, data)
+    return redirect('mostrar_professor')
 
 @login_required
 def listar_grupo(request, template_name='opta_app/lista_grupo.html'):
@@ -23,27 +26,35 @@ def listar_grupo(request, template_name='opta_app/lista_grupo.html'):
 
 @login_required
 def criar_grupo(request, pk_prof, template_name='opta_app/form_grupo.html'):
-    professor = Professor.objects.get(user_id=request.user.id)
-    form = GrupoForm(request.POST or None)
-    if form.is_valid():
-        grupo = form.save()
-        relacionamento = ProfessorGrupo.objects.create(professor_id=professor.id, grupo_id=grupo.pk)
-        return redirect('mostrar_professor')
-    return render(request, template_name, {'form':form})
+    professor = get_object_or_404(Professor, pk=pk_prof)
+    if request.user.professor == professor:
+        form = GrupoForm(request.POST or None)
+        if form.is_valid():
+            grupo = form.save()
+            relacionamento = ProfessorGrupo.objects.create(professor_id=professor.id, grupo_id=grupo.pk)
+            return redirect('mostrar_professor')
+        return render(request, template_name, {'form':form})
+    return redirect('mostrar_professor')
 
 @login_required
 def atualizar_grupo(request, pk_prof, pk_grupo, template_name='opta_app/form_grupo.html'):
-    grupo = get_object_or_404(Grupo, pk=pk_grupo)
-    form = GrupoForm(request.POST or None, instance=grupo)
-    if form.is_valid():
-        form.save()
-        return redirect('mostrar_professor')
-    return render(request, template_name, {'form':form})
+    professor = get_object_or_404(Professor, pk=pk_prof)
+    if request.user.professor == professor:
+        grupo = get_object_or_404(Grupo, pk=pk_grupo)
+        form = GrupoForm(request.POST or None, instance=grupo)
+        if form.is_valid():
+            form.save()
+            return redirect('mostrar_professor')
+        return render(request, template_name, {'form':form})
+    return redirect('mostrar_professor')
 
 @login_required
 def excluir_grupo(request, pk_prof, pk_grupo, template_name='opta_app/grupo_confirm_delete.html'):
-    grupo = get_object_or_404(Grupo, pk=pk_grupo)
-    if request.method=='POST':
-        grupo.delete()
-        return redirect('mostrar_professor')
-    return render(request, template_name, {'object':grupo})
+    professor = get_object_or_404(Professor, pk=pk_prof)
+    if request.user.professor == professor:
+        grupo = get_object_or_404(Grupo, pk=pk_grupo)
+        if request.method=='POST':
+            grupo.delete()
+            return redirect('mostrar_professor')
+        return render(request, template_name, {'object':grupo})
+    return redirect('mostrar_professor')
